utils/slice.go
package utils

func RemoveDuplicates(s []string) []string {
	length := len(s) - 1
	for i := 0; i < length; i++ {
		for j := i + 1; j <= length; j++ {
			if s[i] == s[j] {
				s[j] = s[length]
				s = s[0:length]
				length--
				j--
			}
		}
	}

	return s
}

// DiffArrays will return an array that contains only the keys
// in a that do not appear in b
func DiffArrays(a []string, b []string) []string {
	var diff []string

	for _, s1 := range a {
		found := false
		for _, s2 := range b {
			if s1 == s2 {
				found = true
				break
			}
		}
		// String not found. We add it to return slice
		if !found {
			diff = append(diff, s1)
		}
	}

	return diff
}
